Edit migration for support closure table scheme

+++ migrations/versions/954c8dfb878e_create_user_table.py
@@ -19,7 +19,7 @@ import sqlalchemy as sa
 def upgrade():
     op.create_table(
         'user',
-        sa.Column('id', sa.BigInteger(), nullable=False),
+        sa.Column('id', sa.Integer(), nullable=False),
         sa.Column('name', sa.String(100), unique=True, nullable=False),
         sa.PrimaryKeyConstraint('id')
     )
